requirements.txt
# YOLOv5 requirements
# Usage: pip install -r requirements.txt

gitpython>=3.1.30
matplotlib>=3.3
numpy>=1.22.2
opencv-python>=4.1.1
scikit-learn>=1.3.2
torch>=1.8.0
torchvision>=0.16.1
torchaudio>=2.1.1
tensorflow>=2.4.0
keras>=2.9
pyinstaller>=6.2.0
tkinterdnd2>=0.3.0
